refactor(crossmount): parse_size joins _human_size in formatting utils; combine renamed exit
fanout/combine held only exit-code merging after the wc/du split, so it is now fanout/exit; the TS byte-join helpers move next to their only caller in fanout.ts.
